+++ MyDishesApp.Service/Dtos/Auth/UserDto.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// The user model for authorization
     /// </summary>
-    public class User
+    public class UserDto
     {
         public string Email { get; set; }
         public string FirstName { get; set; }
