--- MyDishesApp.Service/Service/Interfaces/IDishService.cs
@@ -1,6 +1,0 @@
-﻿namespace MyDishesApp.Service.Service.Interfaces
-{
-    public interface IDishService
-    {
-    }
-}
