+++ MyDishesApp.Service/Services/Interfaces/IDishService.cs
@@ -1,0 +1,11 @@
+﻿using MyDishesApp.Service.Dtos;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace MyDishesApp.Service.Service.Interfaces
+{
+    public interface IDishService
+    {
+        Task<IEnumerable<DishDto>> GetDishListAsync();
+    }
+}
