+++ MyDishesApp.WebApi/Controllers/IngredientController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using MyDishesApp.Repository.Repositories.Interfaces;
+using MyDishesApp.Service.Dtos;
 using System;
 using System.Collections.Generic;
 using System.Linq;
